+++ sdk/communication/communication-phone-numbers/src/models.ts
@@ -3,6 +3,7 @@
 
 import { OperationOptions } from "@azure/core-client";
 import {
+  OperatorInformationRequestOptions,
   PhoneNumberAssignmentType,
   PhoneNumberSearchRequest,
   PhoneNumbersListAreaCodesOptionalParams,
@@ -69,7 +70,9 @@ export interface ListLocalitiesOptions extends OperationOptions {
 /**
  * Additional options for the search operator information request.
  */
-export interface SearchOperatorInformationOptions extends OperationOptions {}
+export interface SearchOperatorInformationOptions extends OperationOptions {
+  options?: OperatorInformationRequestOptions[];
+}
 
 /**
  * Additional options that can be passed to list SIP routes.
@@ -110,6 +113,8 @@ export {
   OperatorInformation,
   OperatorInformationResult,
   OperatorNumberType,
+  OperatorInformationRequestOptions,
+  KnownOperatorInformationRequestOptions,
 } from "./generated/src/models/";
 
 export { SipRoutingError, SipTrunkRoute } from "./generated/src/siprouting/models";