+++ sdk/communication/communication-phone-numbers/test/public/numberLookup.spec.ts
@@ -7,6 +7,7 @@ import { Context } from "mocha";
 import { PhoneNumbersClient } from "../../src";
 import { createRecordedClient } from "./utils/recordedClient";
 import { getPhoneNumber } from "./utils/testPhoneNumber";
+import { KnownOperatorInformationRequestOptions } from "../../src/generated/src";
 
 describe(`PhoneNumbersClient - look up phone number`, function () {
   let recorder: Recorder;
@@ -40,5 +41,33 @@ describe(`PhoneNumbersClient - look up phone number`, function () {
       assert.strictEqual(error.code, "BadRequest");
       assert.strictEqual(error.message, "Can only accept one phoneNumber");
     }
-  });
+  }).timeout(60000);
+
+  it("respects IncludeOperatorDetails option", async function (this: Context) {
+    const phoneNumbers = [getPhoneNumber()];
+    let operatorInformation = await client.searchOperatorInformation(phoneNumbers);
+
+    let resultPhoneNumber = operatorInformation.values
+      ? operatorInformation.values[0].phoneNumber
+      : "";
+    assert.strictEqual(resultPhoneNumber, phoneNumbers[0]);
+    assert.isNotNull(operatorInformation.values ? operatorInformation.values[0].nationalFormat : null);
+    assert.isNotNull(operatorInformation.values ? operatorInformation.values[0].internationalFormat : null);
+    assert.isNull(operatorInformation.values ? operatorInformation.values[0].isoCountryCode : null);
+    assert.isNull(operatorInformation.values ? operatorInformation.values[0].numberType : null);
+    assert.isNull(operatorInformation.values ? operatorInformation.values[0].operatorDetails : null);
+
+    operatorInformation = await client.searchOperatorInformation(phoneNumbers, { "options": [ KnownOperatorInformationRequestOptions.IncludeOperatorDetails ]});
+
+    resultPhoneNumber = operatorInformation.values
+      ? operatorInformation.values[0].phoneNumber
+      : "";
+    assert.strictEqual(resultPhoneNumber, phoneNumbers[0]);
+    assert.isNotNull(operatorInformation.values ? operatorInformation.values[0].nationalFormat : null);
+    assert.isNotNull(operatorInformation.values ? operatorInformation.values[0].internationalFormat : null);
+    assert.isNotNull(operatorInformation.values ? operatorInformation.values[0].isoCountryCode : null);
+    assert.isNotNull(operatorInformation.values ? operatorInformation.values[0].numberType : null);
+    assert.isNotNull(operatorInformation.values ? operatorInformation.values[0].operatorDetails : null);
+
+  }).timeout(60000);
 });