+++ sdk/communication/communication-phone-numbers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 1.3.0-beta.2 (2023-11-08)
+
+### Features Added
+
+- Update Number Lookup Preview
+- Add format fields to response
+- Add support for a format only search
+
 ## 1.3.0-beta.1 (2023-07-21)
 
 ### Features Added